tests/wpt/tests/css/css-align/parsing/row-gap-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Box Alignment Level 3: parsing row-gap with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-row-gap">
<meta name="assert" content="row-gap supports only the grammar 'normal | <length-percentage>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("row-gap", "auto");
test_invalid_value("row-gap", "-10px");

test_invalid_value("row-gap", "10px 20%");
test_invalid_value("row-gap", "normal 10px");
</script>
</body>
</html>
